% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famnet-package.R
\docType{package}
\name{famnet-package}
\alias{famnet}
\alias{famnet-package}
\title{famnet: Characterizing a Family Network}
\description{
Calculates relations between a family network, based on an individual ID,
mother ID, and father ID.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lwjohnst86/famnet}
  \item Report bugs at \url{https://github.com/lwjohnst86/famnet/issues}
}

}
\author{
\strong{Maintainer}: Luke Johnston \email{lwjohnst@gmail.com}

Authors:
\itemize{
  \item Omar Silverman
  \item Adam Hulman
}

Other contributors:
\itemize{
  \item Daniel Witte [contributor]
}

}
\keyword{internal}
