% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/famnet-package.R
\docType{data}
\name{fmn_df}
\alias{fmn_df}
\title{IDs for persons and their mother and/or father.}
\format{Dataframe with 50 rows and 4 variables:
\itemize{
\item PersonID: Person's ID
\item MotherID: Person's mother's ID
\item PersonID: Person's father's ID
\item Sex: Sex of the person
}}
\usage{
fmn_df
}
\description{
A dataset that contains a person's ID, as well as ID's for their mother
and/or father. This dataset is used to show how the famnet functions work.
}
\keyword{datasets}
