% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siblings.R
\name{fmn_siblings}
\alias{fmn_siblings}
\alias{fmn_siblings_tv}
\alias{fmn_siblings_dt}
\title{Determine who are siblings using only ID, mother ID, and father ID}
\usage{
fmn_siblings(.data, .personid, .motherid, .fatherid)

fmn_siblings_tv(.data, .personid, .motherid, .fatherid)

fmn_siblings_dt(.data, .personid, .motherid, .fatherid)
}
\arguments{
\item{.data}{Data that contains columns of persons IDs, for the individual,
mother, and father.}

\item{.personid}{Column with main person's ID.}

\item{.motherid}{Column with person's mother's ID.}

\item{.fatherid}{Column with person's father's ID.}
}
\value{
A dataset with at least:
\itemize{
\item \code{PersonID}: An individual's ID
\item \code{RelativeID}: An ID of one of the individual's relative or family member
\item \code{RelativeType}: The type of relationship between the person and the relative
}
}
\description{
This function is best used with data on a very large population, that has
information on an individual as well as that person's parent's IDs. Common
datasets with this information would be population level register databases,
such as found in Denmark or Sweden.
}
\section{Functions}{
\itemize{
\item \code{fmn_siblings_tv}: Function to determine sibling relationships, using
"\href{https://www.tidyverse.org/packages/}{tidyverse}" functions (e.g.
\href{https://dplyr.tidyverse.org/}{dplyr}).

\item \code{fmn_siblings_dt}: Function to determine sibling relationships using
\href{https://github.com/Rdatatable/data.table/wiki}{data.table} functions.
}}

\examples{

# Uses tidyverse under the hood. There is a data.table version available.
fmn_siblings(fmn_df, "PersonID", "MotherID", "FatherID")
}
