% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents.R
\name{fmn_mother_tv}
\alias{fmn_mother_tv}
\alias{fmn_father_tv}
\title{Determine the mother relationship.}
\usage{
fmn_mother_tv(.data, .personid, .motherid)

fmn_father_tv(.data, .personid, .fatherid)
}
\arguments{
\item{.data}{Dataset with person and mother id, at least.}

\item{.personid}{Individual's ID.}

\item{.motherid}{Individual's mother's ID.}

\item{.fatherid}{Individual's father's ID.}
}
\value{
Dataset with the columns: person ID, RelativeID, RelativeType.
}
\description{
Fairly simple wrapper around an \code{\link[dplyr:if_else]{dplyr::if_else()}} condition. This function
is a tidyverse (\code{_tv}) version. The goal of the output data is to mimic a
network (start node, end node, edge type).
}
\examples{

fmn_mother_tv(fmn_df, "PersonID", "MotherID")
fmn_father_tv(fmn_df, "PersonID", "FatherID")

}
